+++ src/resources/boards/board.router.js
@@ -1,49 +1,10 @@
 const router = require('express').Router();
-const boardService = require('./board.service');
+const controller = require('./board.controller');
 
-router.route('/').get(async (req, res, next) => {
-  try {
-    const board = await boardService.getAll();
-    res.json(board);
-  } catch(err) {
-    next(err);
-  }
-});
-
-router.route('/:id').get(async (req, res, next) => {
-  try {
-    const board = await boardService.getById(req.params.id);
-    res.json(board);
-  } catch(err) {
-    next(err);
-  }
-});
-
-router.route('/').post(async (req, res, next) => {
-  try {
-    const board = await boardService.createBoard(req.body);
-    res.status(201).json(board);
-  } catch(err) {
-    next(err);
-  }
-});
-
-router.route('/:id').put(async (req, res, next) => {
-  try {
-    const board = await boardService.updateBoard(req.params.id, req.body);
-    res.json(board);
-  } catch(err) {
-    next(err);
-  }
-});
-
-router.route('/:id').delete(async (req, res, next) => {
-  try {
-    await boardService.deleteBoard(req.params.id);
-    res.sendStatus(200);
-  } catch(err) {
-    next(err);
-  }
-});
+router.route('/').get(controller.getAll);
+router.route('/:id').get(controller.getById);
+router.route('/').post(controller.createBoard);
+router.route('/:id').put(controller.updateBoard);
+router.route('/:id').delete(controller.deleteBoard);
 
 module.exports = router;
